pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt-sovits-onnx"
version = "0.1.0"
authors = [
  { name="High_Logic", email="highlogic233@gmail.com" },
]
description = "An ONNX-based inference engine for the GPT-Sovits project, featuring original model conversion support."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT License" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "onnx",
    "onnxruntime",
    "numpy",
    "pyopenjtalk",
    "soundfile",
    "soxr",
    "pyyaml",
    "rich",
    "pyaudio",
]

[project.urls]

# Homepage = "https://github.com/your-username/your-repo"
# Bug Tracker = "https://github.com/your-username/your-repo/issues"